+++ sdk/spring/spring-cloud-azure-stream-binder-eventhubs/src/main/java/com/azure/spring/cloud/stream/binder/eventhubs/core/provisioning/EventHubsChannelResourceManagerProvisioner.java
@@ -3,7 +3,7 @@
 
 package com.azure.spring.cloud.stream.binder.eventhubs.core.provisioning;
 
-import com.azure.spring.resourcemanager.provisioning.EventHubsProvisioner;
+import com.azure.spring.cloud.resourcemanager.provisioning.EventHubsProvisioner;
 import org.springframework.lang.NonNull;
 import org.springframework.util.Assert;