+++ sdk/spring/spring-cloud-azure-stream-binder-servicebus/src/main/java/com/azure/spring/cloud/stream/binder/servicebus/core/provisioning/ServiceBusChannelResourceManagerProvisioner.java
@@ -3,7 +3,7 @@
 
 package com.azure.spring.cloud.stream.binder.servicebus.core.provisioning;
 
-import com.azure.spring.resourcemanager.provisioning.ServiceBusProvisioner;
+import com.azure.spring.cloud.resourcemanager.provisioning.ServiceBusProvisioner;
 import com.azure.spring.cloud.service.servicebus.properties.ServiceBusEntityType;
 import org.springframework.lang.NonNull;
 import org.springframework.util.Assert;